requirements.txt
tkinter
requests>=2.31.0
sqlite3
gitpython>=3.1.40
openai>=1.3.0
Pillow>=10.0.0
customtkinter>=5.2.0
anthropic>=0.7.0
python-dotenv>=1.0.0
psutil>=5.9.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0 